asset rebuild and build warning fix

+++ MagpieGameMode.cpp
@@ -321,7 +321,7 @@ namespace Magpie {
                     float window_x_anchor = ((x_anchor + aspect) / (2*aspect)) * viewport[2];
                     float window_y_anchor = ((y_anchor + 1) / 2) * viewport[3];
 
-                    uint32_t escape_time_remaining = std::ceil(game.escape_timer - game.elapsed_in_escape);
+                    uint32_t escape_time_remaining = (uint32_t)std::ceil(game.escape_timer - game.elapsed_in_escape);
                     RenderText(tutorial_font.value, "ESCAPE TO THE BACK! " + std::to_string(escape_time_remaining),
                             window_x_anchor, window_y_anchor, font_scale, glm::vec3(1.0f, 0.2f, 0.2f));
                 }
